Conflicts:
      doc/errata/staging/univention-ad-connector.yaml

were fixed.

+++ management/univention-management-console/src/univention/management/console/protocol/server.py
@@ -36,6 +36,7 @@ Defines the basic class for an UMC server.
 """
 
 import os
+import errno
 import fcntl
 import socket
 import resource
@@ -101,7 +102,7 @@ class MagicBucket(object):
 			CORE.info('Shutting down connection %s' % sock)
 			self._cleanup(sock)
 
-	def _receive(self, socket):
+	def _receive(self, sock):
 		"""Signal callback: Handles incoming data. Processes SSL events
 		and parses the incoming data. If a valid UMCP was found it is
 		passed to _handle.
@@ -111,7 +112,7 @@ class MagicBucket(object):
 		data = ''
 
 		try:
-			data = socket.recv(RECV_BUFFER_SIZE)
+			data = sock.recv(RECV_BUFFER_SIZE)
 		except SSL.WantReadError:
 			# this error can be ignored (SSL need to do something)
 			return True
@@ -120,15 +121,19 @@ class MagicBucket(object):
 				CRYPT.warn('The socket was closed by the client.')
 			else:
 				CRYPT.error('SSL error in _receive: %s.' % (exc,))
-			self._cleanup(socket)
+			self._cleanup(sock)
+			return False
+		except socket.error as exc:
+			CORE.warn('Socket error in _receive: %s. Probably close (114).' % (exc,))
+			self._cleanup(sock)
 			return False
 
 		if not data:
-			self._cleanup(socket)
+			self._cleanup(sock)
 			return False
 
 		try:
-			state = self.__states[socket]
+			state = self.__states[sock]
 		except KeyError:
 			return False
 		state.buffer += data
@@ -149,16 +154,16 @@ class MagicBucket(object):
 			CORE.process('Parse error: %r' % (exc,))
 			if msg.id is None:
 				# close the connection in case we use could not parse the header
-				self._cleanup(socket)
+				self._cleanup(sock)
 				return False
 			state.requests[msg.id] = msg
 			state.session.execute('parse_error', msg, exc)
 
 		return True
 
-	def _do_send(self, socket):
+	def _do_send(self, sock):
 		try:
-			state = self.__states[socket]
+			state = self.__states[sock]
 		except KeyError:
 			CORE.warn('The socket was already removed.')
 			return False
@@ -168,7 +173,7 @@ class MagicBucket(object):
 			CORE.error('The response queue for %r is empty.' % (state,))
 			return False
 		try:
-			ret = socket.send(first)
+			ret = sock.send(first)
 			if ret < len(first):
 				state.resend_queue.insert(0, (id, first[ret:]))
 			else:
@@ -180,7 +185,11 @@ class MagicBucket(object):
 			return True
 		except (SSL.SysCallError, SSL.Error) as error:
 			CRYPT.warn('SSL error in _do_send: %s. Probably the socket was closed by the client.' % str(error))
-			self._cleanup(socket)
+			self._cleanup(sock)
+			return False
+		except socket.error as exc:
+			CORE.warn('socket.error in _do_send: %s. Probably the socket was closed by the client.' % (exc,))
+			self._cleanup(sock)
 			return False
 
 		return (len(state.resend_queue) > 0)
@@ -220,7 +229,10 @@ class MagicBucket(object):
 		except (SSL.SysCallError, SSL.Error, socket.error) as error:
 			CRYPT.warn('SSL error in _response: %s. Probably the socket was closed by the client.' % str(error))
 			self._cleanup(state.socket)
-		except:  # close the connection to the client. we can't do anything else
+		except socket.error as exc:
+			CORE.warn('socket error in _response: %s. Probably the socket was closed by the client.' % (exc,))
+			self._cleanup(state.socket)
+		except Exception:  # close the connection to the client. we can't do anything else
 			CORE.error('FATAL ERROR: %s' % (traceback.format_exc(),))
 			self._cleanup(state.socket)
 
@@ -266,23 +278,28 @@ class Server(signals.Provider):
 		CORE.info('Initialising server process')
 		self.__port = port
 		self.__unix = unix
+		self.__realtcpsocket = None
+		self.__realunixsocket = None
 		self.__ssl = ssl
 		if self.__unix:
 			CORE.info('Using a UNIX socket')
-			self.__realsocket = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-		else:
+			self.__realunixsocket = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+		if self.__port:
 			CORE.info('Using a TCP socket')
 			try:
-				self.__realsocket = socket.socket(socket.AF_INET6, socket.SOCK_STREAM)
-			except:
+				self.__realtcpsocket = socket.socket(socket.AF_INET6, socket.SOCK_STREAM)
+			except Exception:
 				CORE.warn('Cannot open socket with AF_INET6 (Python reports socket.has_ipv6 is %s), trying AF_INET' % socket.has_ipv6)
-				self.__realsocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+				self.__realtcpsocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
-		self.__realsocket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-		self.__realsocket.setblocking(0)
-		fcntl.fcntl(self.__realsocket.fileno(), fcntl.F_SETFD, 1)
+		for sock in (self.__realtcpsocket, self.__realunixsocket):
+			if sock is None:
+				continue
+			sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+			sock.setblocking(0)
+			fcntl.fcntl(sock.fileno(), fcntl.F_SETFD, 1)
 
-		if self.__ssl and not self.__unix:
+		if self.__ssl and self.__port:
 			CORE.info('Setting up SSL configuration')
 			self.crypto_context = SSL.Context(SSL.TLSv1_METHOD)
 			self.crypto_context.set_cipher_list(ucr.get('umc/server/ssl/ciphers', 'DEFAULT'))
@@ -300,32 +317,36 @@ class Server(signals.Provider):
 				CRYPT.warn('Communication will not be encrypted!')
 				self.__ssl = False
 				self.crypto_context = None
-				self.__realsocket.bind(('', self.__port))
+				self.__realtcpsocket.bind(('', self.__port))
 				CRYPT.info('Server listening to unencrypted connections')
-				self.__realsocket.listen(SERVER_MAX_CONNECTIONS)
+				self.__realtcpsocket.listen(SERVER_MAX_CONNECTIONS)
 
 			if self.crypto_context:
-				self.connection = SSL.Connection(self.crypto_context, self.__realsocket)
+				self.connection = SSL.Connection(self.crypto_context, self.__realtcpsocket)
 				self.connection.setblocking(0)
 				self.connection.bind(('', self.__port))
 				self.connection.set_accept_state()
 				CRYPT.info('Server listening to SSL connections')
 				self.connection.listen(SERVER_MAX_CONNECTIONS)
-		else:
+		elif not self.__ssl and self.__port:
 			self.crypto_context = None
-			if self.__unix:
-				try:
-					# ensure that the UNIX socket is only accessible by root
-					old_umask = os.umask(0o077)
-					self.__realsocket.bind(self.__unix)
-					# restore old umask
-					os.umask(old_umask)
-				except EnvironmentError:
+			self.__realtcpsocket.bind(('', self.__port))
+			CRYPT.info('Server listening to TCP connections')
+			self.__realtcpsocket.listen(SERVER_MAX_CONNECTIONS)
+
+		if self.__unix:
+			# ensure that the UNIX socket is only accessible by root
+			old_umask = os.umask(0o077)
+			try:
+				self.__realunixsocket.bind(self.__unix)
+			except EnvironmentError:
+				if os.path.exists(self.__unix):
 					os.unlink(self.__unix)
-			else:
-				self.__realsocket.bind(('', self.__port))
-			CRYPT.info('Server listening to connections')
-			self.__realsocket.listen(SERVER_MAX_CONNECTIONS)
+			finally:
+				# restore old umask
+				os.umask(old_umask)
+			CRYPT.info('Server listening to UNIX connections')
+			self.__realunixsocket.listen(SERVER_MAX_CONNECTIONS)
 
 		self.__magic = magic
 		self.__magicClass = magicClass
@@ -335,10 +356,12 @@ class Server(signals.Provider):
 		else:
 			self.signal_new('session_new')
 
-		if self.__ssl and not self.__unix:
+		if self.__ssl:
 			notifier.socket_add(self.connection, self._connection)
-		else:
-			notifier.socket_add(self.__realsocket, self._connection)
+		if (not self.__ssl and self.__port):
+			notifier.socket_add(self.__realtcpsocket, self._connection)
+		if self.__unix:
+			notifier.socket_add(self.__realunixsocket, self._connection)
 
 	def __verify_cert_cb(self, conn, cert, errnum, depth, ok):
 		CORE.info('__verify_cert_cb: Got certificate: %s' % cert.get_subject())
@@ -351,16 +374,29 @@ class Server(signals.Provider):
 		try:
 			socket, addr = socket.accept()
 		except EnvironmentError as exc:
+			if exc.errno == errno.EAGAIN:
+				# got an EAGAIN --> try again later
+				return True
 			CORE.error('Cannot accept new connection: %s' % (exc,))
-			soft, hard = resource.getrlimit(resource.RLIMIT_NOFILE)
-			resource.setrlimit(resource.RLIMIT_NOFILE, (soft + 2, hard + 2))
-			try:
-				socket, addr = socket.accept()
-				socket.close()
-			except EnvironmentError:
-				pass
-			finally:
-				resource.setrlimit(resource.RLIMIT_NOFILE, (soft, hard))
+			if exc.errno == errno.EMFILE:
+				# got an EMFILE --> Too many open files
+				# If the process permanently lacks free file descriptors, incoming
+				# connections waiting in the listening socket backlog will starve.
+				# Therefore the limit is temporarily increased by 2 and the connection
+				# waiting in the backlog is temporarily accepted and immediately
+				# closed again to provoke an error message in the user's browser.
+				soft, hard = resource.getrlimit(resource.RLIMIT_NOFILE)
+				resource.setrlimit(resource.RLIMIT_NOFILE, (soft + 2, hard + 2))
+				try:
+					socket, addr = socket.accept()
+					socket.close()
+				except EnvironmentError:
+					pass
+				finally:
+					resource.setrlimit(resource.RLIMIT_NOFILE, (soft, hard))
+			else:
+				# unknown errno - log traceback and continue
+				CORE.error(traceback.format_exc())
 			return True
 		socket.setblocking(0)
 		if addr:
@@ -381,15 +417,20 @@ class Server(signals.Provider):
 		if self.__bucket:
 			self.__bucket.exit()
 
-		if self.__ssl and not self.__unix:
+		if self.__ssl and self.__port:
 			notifier.socket_remove(self.connection)
 			self.connection.close()
-		elif self.__realsocket:
-			notifier.socket_remove(self.__realsocket)
-			self.__realsocket.close()
-			self.__realsocket = None
+		elif not self.__ssl and self.__port and self.__realtcpsocket:
+			notifier.socket_remove(self.__realtcpsocket)
+			self.__realtcpsocket.close()
+			self.__realtcpsocket = None
 		if self.__unix:
-			os.unlink(self.__unix)
+			if self.__realunixsocket is not None:
+				notifier.socket_remove(self.__realunixsocket)
+				self.__realunixsocket.close()
+				self.__realunixsocket = None
+			if os.path.exists(self.__unix):
+				os.unlink(self.__unix)
 			self.__unix = None
 
 		self.__bucket = None
